sdk/aptos: move shared methods into baseapi
