+++ sdk/js/src/aptos/api/base.ts
@@ -1,0 +1,45 @@
+import { AptosAccount } from "aptos";
+import { AptosClientWrapper } from "../client";
+
+export class AptosBaseApi {
+  protected client: AptosClientWrapper;
+  protected address: string;
+
+  constructor(client: AptosClientWrapper, network: string) {
+    this.client = client;
+    this.address = "";
+  }
+
+  // Contract upgrade
+
+  authorizeUpgrade = (sender: AptosAccount, vaa: Uint8Array) => {
+    const payload = {
+      function: `${this.address}::contract_upgrade::submit_vaa`,
+      type_arguments: [],
+      arguments: [vaa],
+    };
+    return this.client.executeEntryFunction(sender, payload);
+  };
+
+  upgradeContract = (
+    sender: AptosAccount,
+    metadataSerialized: Uint8Array,
+    code: Array<Uint8Array>,
+  ) => {
+    const payload = {
+      function: `${this.address}::contract_upgrade::upgrade`,
+      type_arguments: [],
+      arguments: [metadataSerialized, code],
+    };
+    return this.client.executeEntryFunction(sender, payload);
+  };
+
+  migrateContract = (sender: AptosAccount) => {
+    const payload = {
+      function: `${this.address}::contract_upgrade::migrate`,
+      type_arguments: [],
+      arguments: [],
+    };
+    return this.client.executeEntryFunction(sender, payload);
+  };
+}
