+++ sdk/js/src/aptos/api/tokenBridge.ts
@@ -2,14 +2,11 @@ import { AptosAccount } from "aptos";
 import { assertChain } from "../../utils";
 import { AptosClientWrapper } from "../client";
 import { deriveWrappedAssetAddress } from "../utils";
+import { AptosBaseApi } from "./base";
 
-export class AptosTokenBridgeApi {
-  private client: AptosClientWrapper;
-  private address: string;
-
+export class AptosTokenBridgeApi extends AptosBaseApi {
   constructor(client: AptosClientWrapper, network: string) {
-    this.client = client;
-    this.address = "";
+    super(client, network);
   }
 
   // Complete transfer
@@ -43,39 +40,6 @@ export class AptosTokenBridgeApi {
       function: `${this.address}::complete_transfer_with_payload::submit_vaa`,
       type_arguments: [`${assetContract}::coin::T`],
       arguments: [vaa],
-    };
-    return this.client.executeEntryFunction(sender, payload);
-  };
-
-  // Contract upgrade
-
-  authorizeUpgrade = (sender: AptosAccount, vaa: Uint8Array) => {
-    const payload = {
-      function: `${this.address}::contract_upgrade::submit_vaa`,
-      type_arguments: [],
-      arguments: [vaa],
-    };
-    return this.client.executeEntryFunction(sender, payload);
-  };
-
-  upgradeContract = (
-    sender: AptosAccount,
-    metadataSerialized: Uint8Array,
-    code: Array<Uint8Array>,
-  ) => {
-    const payload = {
-      function: `${this.address}::contract_upgrade::upgrade`,
-      type_arguments: [],
-      arguments: [metadataSerialized, code],
-    };
-    return this.client.executeEntryFunction(sender, payload);
-  };
-
-  migrateContract = (sender: AptosAccount) => {
-    const payload = {
-      function: `${this.address}::contract_upgrade::migrate`,
-      type_arguments: [],
-      arguments: [],
     };
     return this.client.executeEntryFunction(sender, payload);
   };
