+++ sdk/js/src/aptos/api/wormhole.ts
@@ -1,47 +1,11 @@
-import { AptosAccount } from 'aptos';
-import { AptosClientWrapper } from '../client';
+import { AptosAccount } from "aptos";
+import { AptosClientWrapper } from "../client";
+import { AptosBaseApi } from "./base";
 
-export class AptosWormholeApi {
-  private client: AptosClientWrapper;
-  private address: string;
-
+export class AptosWormholeApi extends AptosBaseApi {
   constructor(client: AptosClientWrapper, network: string) {
-    this.client = client;
-    this.address = '';
+    super(client, network);
   }
-
-  // Contract upgrade
-
-  authorizeUpgrade = (sender: AptosAccount, vaa: Uint8Array) => {
-    const payload = {
-      function: `${this.address}::contract_upgrade::submit_vaa`,
-      type_arguments: [],
-      arguments: [vaa],
-    };
-    return this.client.executeEntryFunction(sender, payload);
-  };
-
-  upgradeContract = (
-    sender: AptosAccount,
-    metadataSerialized: Uint8Array,
-    code: Array<Uint8Array>,
-  ) => {
-    const payload = {
-      function: `${this.address}::contract_upgrade::upgrade`,
-      type_arguments: [],
-      arguments: [metadataSerialized, code],
-    };
-    return this.client.executeEntryFunction(sender, payload);
-  };
-
-  migrateContract = (sender: AptosAccount) => {
-    const payload = {
-      function: `${this.address}::contract_upgrade::migrate`,
-      type_arguments: [],
-      arguments: [],
-    };
-    return this.client.executeEntryFunction(sender, payload);
-  };
 
   // Guardian set upgrade
